+++ react-project/src/components/About.js
@@ -1,0 +1,26 @@
+import {useLocation, Link} from "react-router-dom";
+import ProductList from "./ProductList";
+
+
+import React from 'react';
+import { AppBar, Tabs, Tab } from '@material-ui/core';
+import Header from './Header';
+
+
+
+
+export const About = () => (
+            <>
+            <Header />
+             <div>
+                 <h1>About</h1>
+                 <p>About, introducing page</p>
+             </div>
+             </>
+         );
+
+
+
+
+
+
