+++ react-project/src/components/Carousel.js
@@ -24,7 +24,7 @@ const Carousel = () => {
 
   const items = [
     {
-      name: '컴퓨터 견적을 위해 도와드립니다.',
+      name: '컴퓨터 견적을 맞추는 것에 어려움을 겪는분은 다음 버튼을 눌러 몇가지 질문에 답해주세요',
     },
     {
       name: '컴퓨터를 사용하는 목적이 무엇인가요?',
