+++ react-project/src/components/Result.js
@@ -1,0 +1,27 @@
+import {useLocation, Link} from "react-router-dom";
+import ProductList from "./ProductList";
+
+
+import React from 'react';
+import { AppBar, Tabs, Tab } from '@material-ui/core';
+import Header from './Header';
+
+
+export const Result = () => (
+    <>
+            <Header />
+    <div>
+        <h1>견적</h1>
+        <p>page for 견적</p>
+
+    </div>
+    </>
+);
+
+
+
+
+
+
+
+
